+++ src/MultiTranslator.php
@@ -45,7 +45,7 @@ use yii\log\Logger;
  */
 class MultiTranslator extends Plugin
 {
-    public string $schemaVersion = '1.3.0';
+    public string $schemaVersion = '1.4.0';
     public bool $hasCpSettings = true;
     public bool $hasCpSection = true;
     public ?string $name = 'Multi Translator';
@@ -146,6 +146,7 @@ class MultiTranslator extends Plugin
             function (RegisterUrlRulesEvent $event) {
                 $event->rules['multi-translator/glossaries/edit/<id:\d+>'] = 'multi-translator/glossaries/edit';
                 $event->rules['multi-translator/glossaries/new'] = 'multi-translator/glossaries/new';
+                $event->rules['multi-translator/glossaries/fetch'] = 'multi-translator/glossaries/fetch';
                 $event->rules['multi-translator/translate/review'] = 'multi-translator/translate/review';
             }
         );
@@ -259,6 +260,7 @@ class MultiTranslator extends Plugin
             Entry::class,
             Asset::class,
             'craft\commerce\elements\Product',
+            'craft\commerce\elements\Variant',
         ];
 
         $existing = [];